Fork Presto ES connector for continued ES 6 support
Upgrading the Presto Elasticsearch connector to use Elasticsearch 7.0
libraries will introduce breaking changes for supporting Elasticsearch 6.0
servers due to the lack of backward compatibility in the Elasticsearch
v7.0 RestHighLevelClient (HLC) and the Types deprecated in APIs in
ES 7.0 introducing breaking changes to the index creation, put mapping,
get mapping, put template, get template and get field mappings APIs.

By forking and creating a presto-elasticsearch-6 version of the
Elasticsearch connector as a new module we can mitigate these
breaking changes so users can stay using the older connector code if
they need to stay on the older ES 6.x servers. As it's own module we can
release this independently and provide instructions for users to retain
ES6 access. If it goes as a submodule in the main Presto repository it can
cause issues with CI and the versioning won't make sense since it would
follow Presto's versioning, when as a standalone plugin it should be
versioned independently.

The existing presto-elasticsearch module source, tests, and config files
were copied, renamed to presto-elasticsearch-6 and copied thi new
repository. Update the name of the plugin to ElasticsearchLegacyPlugin
and the connector name to elasticsearch_legacy to avoid conflicts.